data: Change the airdates for The Cage
This is a bit of a nitpick. There's two entries, 'The Cage' and 'The
Cage (extended)', which is how they're called on the Blu-ray set.
"extended" is a bit of a misnomer and oddity: it includes Gene
Roddenberry's introduction and sign-off from the 1988 broadcast, and
also preserves the mixed black/white and color scenes of the episode,
from when the color portions were believed to be lost. As far as I
know, the episode itself is still 100% identical to the "normal" one.

Anyhow, since the extended one is what broadcast in 1988, let's put
the airdate on that. remastered_airdate still belongs on the first
one, however, since it was shown in complete color and without Gene's
intro and signoff.

FossilOrigin-Name: 378d9c6ee92b8339fb4c4e8a67a0c0c5eaeb34c4a9e4660cef7ce42cb08d8ad2