Remove stardate from Insurrection.
This is actually the only movie (so far) that never declares a stardate.
The previously recorded entry was taken from Memory Alphaʼs “Star Trek
films” article.  It is mainly speculative and based on a rough estimate
of what the stardate should be based on DS9 and Voyager episodes that
aired in December 1998.