Schema: Split episode.date into three columns
Accomidating for Discovery’s third episode, in which the month
and year are known, but not the date.  PostgreSQL doesn’t directly
handle dates where only data is partially known.  I could enter it
as 2256-11-01, but it’s misleading.

Check constraints have been added so that invalid dates (such as
2260-02-30) cannot be entered.  No consideration for leap years is
implemented, so 2259-02-29 would be accepted.

On the ent and dis views, I combine these columns into a proper date
field so date sorting works as expected.  It will only display a value
when all three fields are not null.

Also direct series IDs are now used in all the view/function
definitions.  Saves a lookup, and these IDs will not be changing.